vim-patch:8.1.2319: compiler warning for int size
Problem: Compiler warning for int size.
Solution: Add typecast. (Mike Williams)
https://github.com/vim/vim/commit/07a63d86338476bafbd1a3ec462672df92666498

+++ src/nvim/mouse.c
@@ -410,7 +410,7 @@ bool mouse_comp_pos(win_T *win, int *rowp, int *colp, linenr_T *lnump)
     col += win->w_skipcol;
     // limit to text length plus one
     p = ml_get_buf(win->w_buffer, lnum, false);
-    count = STRLEN(p);
+    count = (int)STRLEN(p);
     if (col > count) {
       col = count;
     }
